+++ apps/web/src/components/Builders/Stories/Hero/assets/base_globe.webp
