+++ apps/web/src/components/Builders/UseCases/assets/BaseAgent.tsx
