+++ apps/web/src/components/Builders/UseCases/assets/Onboarding.tsx
