+++ apps/web/src/components/base-org/shared/TopNavigation/DevelopersDropdown.tsx
@@ -3,12 +3,12 @@ import Title from 'apps/web/src/components/base-org/typography/Title';
 import { TitleLevel } from 'apps/web/src/components/base-org/typography/Title/types';
 import { Icon } from 'apps/web/src/components/Icon/Icon';
 import Link from 'next/link';
-import agentKit from 'apps/web/src/components/Developers/Tools/assets/agentKit.svg';
-import baseNet from 'apps/web/src/components/Developers/Tools/assets/appchains.svg';
-import miniKit from 'apps/web/src/components/Developers/Tools/assets/miniKit.svg';
-import baseWallet from 'apps/web/src/components/Developers/Tools/assets/smartWallet.svg';
-import onchainKit from 'apps/web/src/components/Developers/Tools/assets/onchainKit.svg';
-import verification from 'apps/web/src/components/Developers/Tools/assets/verification.svg';
+import agentKit from 'apps/web/src/components/Builders/Tools/assets/agentKit.svg';
+import baseNet from 'apps/web/src/components/Builders/Tools/assets/appchains.svg';
+import miniKit from 'apps/web/src/components/Builders/Tools/assets/miniKit.svg';
+import baseWallet from 'apps/web/src/components/Builders/Tools/assets/smartWallet.svg';
+import onchainKit from 'apps/web/src/components/Builders/Tools/assets/onchainKit.svg';
+import verification from 'apps/web/src/components/Builders/Tools/assets/verification.svg';
 import Image, { StaticImageData } from 'next/image';
 import options from 'apps/web/src/components/base-org/shared/TopNavigation/assets/developers/options.svg';
 import bug from 'apps/web/src/components/base-org/shared/TopNavigation/assets/developers/bug.svg';
